try some workarounds

+++ vm/message/src/message.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 
 use forest_shim::{message::Message, version::NetworkVersion};
-use fvm::gas::Gas;
+use fvm3::gas::Gas;
 
 /// Semantic validation and validates the message has enough gas.
 pub fn valid_for_block_inclusion(
@@ -36,7 +36,7 @@ pub fn valid_for_block_inclusion(
         anyhow::bail!("gas_limit cannot be greater than block gas limit");
     }
 
-    if Gas::new(msg.gas_limit as i64) < min_gas {
+    if Gas::new(msg.gas_limit) < min_gas {
         anyhow::bail!(
             "gas_limit {} cannot be less than cost {} of storing a message on chain",
             msg.gas_limit,
